Make simple functions inline
