+++ runtime-light/stdlib/fork/fork-context.h
@@ -8,6 +8,7 @@
 #include <optional>
 
 #include "runtime-core/memory-resource/unsynchronized_pool_resource.h"
+#include "runtime-core/utils/kphp-assert-core.h"
 #include "runtime-light/coroutine/task.h"
 #include "runtime-light/stdlib/fork/fork.h"
 #include "runtime-light/utils/concepts.h"
@@ -22,11 +23,25 @@ class ForkComponentContext {
   int64_t next_fork_id_{FORK_ID_INIT};
 
 public:
-  explicit ForkComponentContext(memory_resource::unsynchronized_pool_resource &) noexcept;
+  explicit ForkComponentContext(memory_resource::unsynchronized_pool_resource &memory_resource) noexcept
+    : forks_(unordered_map<int64_t, task_t<fork_result>>::allocator_type{memory_resource}) {}
 
   static ForkComponentContext &get() noexcept;
 
-  int64_t push_fork(task_t<fork_result> &&) noexcept;
+  int64_t push_fork(task_t<fork_result> &&task) noexcept {
+    const auto fork_id{next_fork_id_++};
+    forks_.emplace(fork_id, std::move(task));
+    php_debug("ForkComponentContext: push fork %" PRId64, fork_id);
+    return fork_id;
+  }
 
-  std::optional<task_t<fork_result>> pop_fork(int64_t) noexcept;
+  std::optional<task_t<fork_result>> pop_fork(int64_t fork_id) noexcept {
+    if (const auto it_fork{forks_.find(fork_id)}; it_fork != forks_.cend()) {
+      php_debug("ForkComponentContext: pop fork %" PRId64, fork_id);
+      auto fork{std::move(it_fork->second)};
+      forks_.erase(it_fork);
+      return {std::move(fork)};
+    }
+    return std::nullopt;
+  }
 };
